+++ packages/okam-cli/templates/base/reduxSrc/actions/index.js
@@ -4,4 +4,5 @@
  */
 
 import * as counterActions from './counter';
-export default counterActions;
+export default Object.assign({}, counterActions)
+// export default counterActions; // 这种写法 支付宝小程序编译会出错