+++ packages/okam-cli/templates/base/scripts/ant.config.js
@@ -15,21 +15,6 @@ module.exports = merge({}, require('./base.config'), {
     <% if: ${async} %>
     localPolyfill: [
         'async'
-    ],
+    ]
     <% /if %>
-    processors: {
-        <% if: ${script} === 'babel' %>
-        babel: {
-            options: {
-                presets: ['babel-preset-env']
-            }
-        }
-        <% else %>
-        babel7: {
-            options: {
-                presets: ['@babel/preset-env']
-            }
-        }
-        <% /if %>
-    }
 });